api-python: rename 'velstor.api' to 'velstor.restapi' to make room for high-level apis.
