+++ api-python/velstor/restapi/grid.py
@@ -1,7 +1,7 @@
 import requests
 import json
-from velstor.api.util import urlencode
-from velstor.api.fulfill202 import fulfill202
+from velstor.restapi.util import urlencode
+from velstor.restapi.fulfill202 import fulfill202
 
 #  grid.py:  Operations about grid jobs
 
@@ -10,11 +10,11 @@ def get(session, jobid):
     """Retrieves details about a grid job.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         jobid (str): The job's identifying string.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #
     url = '/'.join([session.base_url(),
@@ -41,11 +41,11 @@ def delete(session, jobid):
     """Deletes information about a grid job.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         jobid (str): The job's identifying string.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #
     url = '/'.join([session.base_url(), 'grid/job', urlencode(jobid)])
@@ -57,13 +57,13 @@ def post(session, jobid, vtrq_id, workspace_name):
     """Posts information about a grid job.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         jobid (str): The job's identifying string.
         vtrq_id (int): The vTRQ of the worksapce name.
         workspace_name (str): The PeerCache workspace employed by the job.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #
     url = '/'.join([session.base_url(), 'grid/job', urlencode(jobid)])
@@ -77,10 +77,10 @@ def list(session):
     """Retrieves a list of existing jobs.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #
     url = '/'.join([session.base_url(), 'grid/jobs'])
