+++ api-python/velstor/restapi/namespace.py
@@ -5,8 +5,8 @@
 #
 import requests
 
-from velstor.api.util import urlencode
-from velstor.api.fulfill202 import fulfill202
+from velstor.restapi.util import urlencode
+from velstor.restapi.fulfill202 import fulfill202
 
 #  namespace.py:  Operations on a TRQ namespace
 
@@ -15,12 +15,12 @@ def consistency_get(session, vtrqid, path):
     """Retrieves the consistency semantics attribute of a namespace node.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified namespace path.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
@@ -39,13 +39,13 @@ def consistency_set(session, vtrqid, value, path):
     """Sets the consistency semantics attribute of a namespace node.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         value (str): Either 'eventual' or 'immediate'
         path (str): Fully-qualified namespace path.
 '
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate value is either immediate or eventual
@@ -65,13 +65,13 @@ def copy_vector(session, vtrqid, pairs, overwrite):
     """Retrieves the consistency semantics attribute of a namespace node.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         pairs (list): A list of (source, destination) pairs.
         overwrite (bool): When true, an existing destination is overwritten with 'cp -r' semantics; otherwise, the pair is skipped.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate pairs is an array
@@ -90,13 +90,13 @@ def delete(session, vtrqid, path, recursive):
     """Removes a namespace node.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified namespace path.
         recursive (bool): When True, a subtree will removed.  Otherwise, only leafs will be removed.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
@@ -113,12 +113,12 @@ def mkdir(session, vtrqid, mode, parents, path):
     """Retrieves the consistency semantics attribute of a namespace node.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified namespace path.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
@@ -137,12 +137,12 @@ def list(session, vtrqid, path):
     """Retrieves the consistency semantics attribute of a namespace node.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified namespace path.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
