+++ api-python/velstor/restapi/service.py
@@ -1,5 +1,5 @@
 import requests
-from velstor.api.fulfill202 import fulfill202
+from velstor.restapi.fulfill202 import fulfill202
 
 # service.py:  Operations on a TRQ
 
@@ -8,11 +8,11 @@ def shutdown(session, vtrqid):
     """Gracefully shuts down a vTRQ.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     # validate vtrqid is an integer
     url = '/'.join([session.base_url()
