+++ api-python/velstor/restapi/vp.py
@@ -1,6 +1,6 @@
 import requests
 
-from velstor.api.fulfill202 import fulfill202
+from velstor.restapi.fulfill202 import fulfill202
 
 #  vp.py:  Operations about VPs
 
@@ -9,12 +9,12 @@ def get(session, vtrqid, vpid):
     """Retrieves details about a mounted client file system.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         vpid (str): An opaque VP (mounted client file system) identifier.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate vpid
@@ -31,12 +31,12 @@ def delete(session, vtrqid, vpid):
     """Dismounts a mounted client file system.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         vpid (str): An opaque VP (mounted client file system) identifier.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate vpid
@@ -53,13 +53,13 @@ def find(session, vtrqid, vp_host, mount_point):
     """Returns a filtered list of VP identifiers.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         vp_host (str): The client hostname to match.
         mount_point (str): The path on the client host at which the VP is mounted.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate vpid
