+++ api-python/velstor/restapi/workspace.py
@@ -3,9 +3,9 @@
 """
 import requests
 import json
-from velstor.api.util import synthetic_response
-from velstor.api.util import urlencode
-from velstor.api.fulfill202 import fulfill202
+from velstor.restapi.util import synthetic_response
+from velstor.restapi.util import urlencode
+from velstor.restapi.fulfill202 import fulfill202
 
 # workspace.py:  Operations on a TRQ's set of workspace specifications
 
@@ -14,12 +14,12 @@ def delete(session, vtrqid, path):
     """Deletes a workspace.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified workspace name.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
@@ -37,12 +37,12 @@ def get(session, vtrqid, path):
     """Retrieves a workspace specification.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified workspace name.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
@@ -62,12 +62,12 @@ def list(session, vtrqid, path):
     The names may represent workspaces or nodes of the hierarchy.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified workspace path.
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
@@ -86,13 +86,13 @@ def set(session, vtrqid, path, spec):
     """Creates or overwrites a workspace.
 
     Args:
-        session (:class:`~velstor.api.session.Session`): Provides security information.
+        session (:class:`~velstor.restapi.session.Session`): Provides security information.
         vtrqid (int): ID of the vTRQ.
         path (str): Fully-qualified workspace name.
         spec (Iterable[str]): Workspace definition
 
     Returns:
-        The return value of :func:`~velstor.api.fulfill202.fulfill202`
+        The return value of :func:`~velstor.restapi.fulfill202.fulfill202`
     """
     #  validate vtrqid is an int
     #  validate path is a string and is absolute
